+++ src/caption/Caption.style.tsx
@@ -6,10 +6,18 @@ export const StyledCaption = styled.div`
   font-size: ${font.s.size};
   line-height: ${font.s.lineHeight};
   color: ${color.lightMidnightGreen};
+  display: flex;
 
-  .kirk-link {
-    font-size: ${font.s.size};
-    line-height: ${font.s.lineHeight};
-    color: ${color.lightMidnightGreen};
+  .kirk-button {
+    display: inline;
+  }
+
+  .kirk-caption-time,
+  .kirk-caption-secondary-content {
+    flex: 1 0 auto;
+  }
+
+  .kirk-caption-secondary-content {
+    text-align: right;
   }
 `