+++ src/_internals/item/Item.tsx
@@ -90,21 +90,6 @@ export const Item = (props: ItemProps) => {
   let Tag = tag.type
   let tagProps = tag.props
 
-  // The code below make sure that if we use a HTML button, a proper type='button' or
-  // type='submit' is added. type='button' is the default.
-  // This type=button is needed to make sure non-submit buttons are not activated when
-  // submitting forms.
-  // See: https://stackoverflow.com/questions/4763638/enter-triggers-button-click
-  let tagTypeProp = {}
-  const isButtonTag = tag.type === 'button' && typeof href !== 'string'
-  if (isButtonTag) {
-    if (tag.props.type) {
-      tagTypeProp = { type: tag.props.type }
-    } else {
-      tagTypeProp = { type: 'button' }
-    }
-  }
-
   if (href) {
     if (typeof href !== 'string') {
       Tag = href.type
@@ -114,6 +99,17 @@ export const Item = (props: ItemProps) => {
       tagProps = { href }
     }
   }
+
+  if (Tag === 'button') {
+    // The code below make sure that if we use a HTML button, a proper type='button' or
+    // type='submit' is added. type='button' is the default.
+    // This type=button is needed to make sure non-submit buttons are not activated when
+    // submitting forms.
+    // See: https://stackoverflow.com/questions/4763638/enter-triggers-button-click
+    const buttonType = tag.props.type || 'button'
+    tagProps = { ...tagProps, type: buttonType }
+  }
+
   const hasRightText = rightTitle || rightBody
 
   const getTextColor = (textColor: string) => (disabled ? color.gray : textColor)
@@ -139,7 +135,6 @@ export const Item = (props: ItemProps) => {
   return (
     <Tag
       {...tagProps}
-      {...tagTypeProp}
       onClick={onClick}
       onFocus={onFocus}
       onBlur={onBlur}