+++ src/_internals/item/Item.unit.tsx
@@ -66,6 +66,17 @@ describe('Item', () => {
     expect(wrapper.find('button[type="button"]').exists()).toBe(true)
   })
 
+  it('Should not add type=button to anchors', () => {
+    // The Item uses a <button> tag and href. Because of the href using an anchor, the item will be
+    // transformed into an HTML anchor, not HTML button.
+    const wrapper = mount(<Item tag={<button />} href={<a href="slug_link" />} />)
+
+    // Verify that there is a proper generated anchor element.
+    expect(wrapper.find('button[type="button"]').exists()).toBe(false)
+    expect(wrapper.find('a[type="button"]').exists()).toBe(false)
+    expect(wrapper.find('a').exists()).toBe(true)
+  })
+
   it('Should use correct button type for submit buttons', () => {
     const wrapper = mount(<Item tag={<button type="submit" />} />)
     expect(wrapper.find('button[type="submit"]').exists()).toBe(true)