+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
 - **[FIX]** Add explicit type"button" to Item using an HTML button to avoid accidental activation on form submit.
-[...]
+- **[FIX]** Make `FilterBar` button not moving when `Loader` is visible
+  [...]
 
 # v40.5.0 (23/09/2020)